divide commands.ts into separates files
