+++ packages/testsuite/cypress/support/resource-utils.ts
@@ -1,3 +1,5 @@
+/// <reference types="cypress" />
+
 Cypress.Commands.add("createMailSession", (managementEndpoint, mailSession) => {
   cy.addAddress(managementEndpoint, ["subsystem", "mail", "mail-session", mailSession.mailSessionName], {
     "jndi-name": mailSession.jndiName,
@@ -10,6 +12,52 @@ Cypress.Commands.add("createOutboundSocketBinding", (managementEndpoint, socketB
     ["socket-binding-group", "standard-sockets", "remote-destination-outbound-socket-binding", socketBinding.name],
     socketBinding
   );
+});
+
+Cypress.Commands.add("addAddress", (managementEndpoint, address, parameters) => {
+  return cy.task("execute:cli", {
+    managementApi: `${managementEndpoint}/management`,
+    operation: "add",
+    address: address,
+    ...parameters,
+  });
+});
+
+Cypress.Commands.add("addAddressIfDoesntExist", (managementEndpoint, address, parameters) => {
+  return cy
+    .task("execute:cli", {
+      managementApi: managementEndpoint + "/management",
+      operation: "validate-address",
+      value: address,
+    })
+    .then((result) => {
+      expect((result as { outcome: string }).outcome).to.equal("success");
+      if (!(result as { result: { valid: boolean } }).result.valid) {
+        cy.task("execute:cli", {
+          managementApi: managementEndpoint + "/management",
+          operation: "add",
+          address: address,
+          ...parameters,
+        });
+      }
+    });
+});
+
+Cypress.Commands.add("removeAddressIfExists", (managementEndpoint, address) => {
+  cy.task("execute:cli", {
+    managementApi: managementEndpoint + "/management",
+    operation: "validate-address",
+    value: address,
+  }).then((result) => {
+    expect((result as { outcome: string }).outcome).to.equal("success");
+    if ((result as { result: { valid: boolean } }).result.valid) {
+      cy.task("execute:cli", {
+        managementApi: managementEndpoint + "/management",
+        operation: "remove",
+        address: address,
+      });
+    }
+  });
 });
 
 export {};
@@ -32,6 +80,9 @@ declare global {
           port: string;
         }
       ): Chainable<void>;
+      addAddress(managementEndpoint: string, address: string[], parameters?: object): Chainable<unknown>;
+      addAddressIfDoesntExist(managementEndpoint: string, address: string[], parameters?: object): Chainable<unknown>;
+      removeAddressIfExists(managementEndpoint: string, address: string[]): void;
     }
   }
 }
